리팩토링 작업
